ARM64: KVM: Only trap PMCR_EL0 access from Guest
We provide direct access of PMU HW to Guest hence we don't trap
and emulate PMU registers.

We also want to reserve one PMU event counter for performance
analysis in EL2-mode. To achieve this we reserve the last event
counter for EL2-mode and fake the number of event counters
available to the Guest by trapping only PMCR_EL0 access.

Signed-off-by: Anup Patel <anup.patel@linaro.org>
Signed-off-by: Pranavkumar Sawargaonkar <pranavkumar@linaro.org>