UBUNTU: [Debian] uploadnum should be the remainder of the version
The uploadnum variable in debian/rules is somewhat confusingly named,
it should be the "remainder of the version after <version>-<abinum>", not
just the uploadnum itself.  This version is only used for identification
in uname -v and there it is completely appropriate for this to be complete
with backport versions etc.

BugLink: http://bugs.launchpad.net/bugs/1407755
Signed-off-by: Andy Whitcroft <apw@canonical.com>