x86, doc: Remove cmdline_size from list of fields to be filled in for EFI handover
The cmdline_size field in setup_header is listed as read-only, and is
never read by the kernel. Therefore, there is no need for it to be set
by a boot loader implementing the EFI handover protocol, so remove it
from the list of fields required to be set by a boot loader.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>
Cc: H. Peter Anvin <hpa@linux.intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org
Signed-off-by: Matt Fleming <matt.fleming@intel.com>
(cherry picked from commit a93e178838b85409601c9b909af5addbbfd90264)
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>