KVM: arm/arm64: Enable the KVM-VFIO device
Original Patch Ref: [RFC,v5,01/13] KVM: arm/arm64: Enable the KVM-VFIO device

The KVM-VFIO device is used by the QEMU VFIO device. It is used to
record the list of in-use VFIO groups so that KVM can manipulate
them. With this series, it will also be used to record the forwarded
IRQs.

Signed-off-by: Kim Phillips <kim.phillips@linaro.org>
Signed-off-by: Eric Auger <eric.auger@linaro.org>
---
This is [PATCH 12/19] KVM: arm/arm64: Enable the KVM-VFIO device